% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruitRanking.R
\name{emptyDF_Builder}
\alias{emptyDF_Builder}
\title{An Auto Empty-Data.Frame Generator}
\usage{
emptyDF_Builder(cols)
}
\arguments{
\item{cols}{the columns which should be set to each column of the empty data.frame}
}
\description{
Build a function which is able to create an empty data.frame
}
\details{
for some cases, we need to initiate an empty data.frame for streaming data in.
}
\author{
Will Kuan <aiien61will@gmail.com>
}

