% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruitRanking.R
\name{grading}
\alias{grading}
\title{Score Each Fruit}
\usage{
grading(age, gender, pregnant)
}
\description{
Calculate score of each fruit based on its price and nutrition.
}
\author{
Will Kuan <aiien61will@gmail.com> Conner Chang
}

