% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_fruitRank}
\alias{run_fruitRank}
\title{Rank fruits from various dimensions.}
\usage{
run_fruitRank()
}
\description{
Rank fruits from various dimensions.
}
\examples{
library(fruitBank)
run_fruitRank()
}

