% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruitRanking.R
\name{grading}
\alias{grading}
\title{Score Each Fruit}
\usage{
<<<<<<< HEAD
grading(age, gender, pregnant, priceTable)
}
\arguments{
\item{priceTable}{a table contains fruit names and each market price.}

\item{nu.df}{nutrition data.frame which is consist of variabls including fruit name, each nutritions.}

\item{rda.df}{Rcommended Dietary Allowance data.frame for each group of different age, gender, and body condition.}
=======
grading(age, gender, pregnant)
>>>>>>> 52ea9c759b14dd82a49e65c9a1fc9fffded6db1a
}
\description{
Calculate score of each fruit based on its price and nutrition.
}
\author{
Will Kuan <aiien61will@gmail.com>, Conner Chang
}

